add effect

+++ Assets/Scripts/Interfaces/IUpgradable.cs
@@ -1,5 +1,5 @@
 public interface IUpgradable
 {
     void Upgrade();
-    bool CanUpgrade(int resources);
+    bool CanUpgrade(float resources);
 }
